+++ spec/factories/lightstreamer/subscription.rb
@@ -4,7 +4,10 @@ FactoryGirl.define do
     fields []
     mode :merge
     adapter nil
+    maximum_update_frequency nil
 
-    initialize_with { new items: items, fields: fields, mode: mode, adapter: adapter }
+    initialize_with do
+      new items: items, fields: fields, mode: mode, adapter: adapter, maximum_update_frequency: maximum_update_frequency
+    end
   end
 end