+++ spec/lightstreamer/cli/commands/stream_command_spec.rb
@@ -21,7 +21,7 @@ describe Lightstreamer::CLI::Main do
     expect(session).to receive(:connect)
 
     expect(Lightstreamer::Subscription).to receive(:new)
-      .with(items: ['item'], fields: ['field'], mode: :merge, adapter: 'adapter')
+      .with(items: ['item'], fields: ['field'], mode: :merge, adapter: 'adapter', maximum_update_frequency: nil)
       .and_return(subscription)
 
     expect(subscription).to receive(:add_data_callback)