+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### 0.4 — Unreleased
 
+- Added support for specifying a subscription's maximum update frequency
 - Added validation of the arguments for control connection requests
 - Fixed handling of `nil` subscription adapters
 - Unhandled exceptions on the internal worker threads now cause the application to terminate